Updating of VIPS Yaml to use EPPO codes representing multiple species instead of individual species codes where the DSS does not differentiate between the
Where the DSS does not differentiate between individual species, it is better, where possible to use an EPPO code that represents a genera. 
This is particularly important for brassicas, where there is a very large number of species grown and the EPPO code conversion will produce a large number ofduplicate entries that the UI or backend has to filter to make unique.

I felt that as there was an EPPO code to represent the brassica genus (BRSSS) that this should be used instead as it simplifies display to the user, removing the need for the user to select from a large number of potential crops.

I have also done the same for the Bremia lactuca model as the DSS does not distinguish between different lettuce species or varieties, replacing these with LACSS.

The aim with the EPPO codes should be to use the minimum number possible to represent the species/varieties that the model can cover.